Rewrote table schemas and added them to README.md.
The two tables, distributors and films, now contain all the columns
found in the spreadsheet.

A significant percentage of the columns in films are either 0 or empty
as they represent the initial values of various global variables. When
a film entry is created, it stands to reason that the gross column would
be 0.

The only additional column to the distributors table is 'attn' which
will likely be empty for most entries.