Updated film name variable and README.md.
The variable name in film.py that stores the given name of a film was
changed to title as that not only is the better term, it prevents
confusion with the name variable in distributor.py.

The README.md file was updated to support proper syntax. Let's hope
it works in this commit.